+++ app/app.go
@@ -389,7 +389,7 @@ func NewAxelarApp(logger log.Logger, db dbm.DB, traceStore io.Writer, loadLatest
 		appCodec, keys[govtypes.StoreKey], app.getSubspace(govtypes.ModuleName), accountK, bankK,
 		&stakingK, govRouter,
 	)
-	govK.SetHooks(govtypes.NewMultiGovHooks(nexusK.Hooks(govK)))
+	govK.SetHooks(govtypes.NewMultiGovHooks(axelarnetK.Hooks(nexusK, govK)))
 
 	// axelar custom keepers
 	// axelarnet / nexus keepers created above
