+++ docs/proto/proto-docs.md
@@ -17,6 +17,7 @@
     - [TokenSent](#axelar.axelarnet.v1beta1.TokenSent)
   
 - [axelar/axelarnet/v1beta1/params.proto](#axelar/axelarnet/v1beta1/params.proto)
+    - [CallContractProposalMinDeposit](#axelar.axelarnet.v1beta1.CallContractProposalMinDeposit)
     - [Params](#axelar.axelarnet.v1beta1.Params)
   
 - [axelar/axelarnet/v1beta1/types.proto](#axelar/axelarnet/v1beta1/types.proto)
@@ -369,7 +370,6 @@
     - [RateLimitUpdated](#axelar.nexus.v1beta1.RateLimitUpdated)
   
 - [axelar/nexus/v1beta1/params.proto](#axelar/nexus/v1beta1/params.proto)
-    - [CallContractProposalMinDeposit](#axelar.nexus.v1beta1.CallContractProposalMinDeposit)
     - [Params](#axelar.nexus.v1beta1.Params)
   
 - [axelar/nexus/v1beta1/genesis.proto](#axelar/nexus/v1beta1/genesis.proto)
@@ -779,6 +779,23 @@
 
 
 
+<a name="axelar.axelarnet.v1beta1.CallContractProposalMinDeposit"></a>
+
+### CallContractProposalMinDeposit
+
+
+
+| Field | Type | Label | Description |
+| ----- | ---- | ----- | ----------- |
+| `chain` | [string](#string) |  |  |
+| `contract_address` | [string](#string) |  |  |
+| `min_deposits` | [cosmos.base.v1beta1.Coin](#cosmos.base.v1beta1.Coin) | repeated |  |
+
+
+
+
+
+
 <a name="axelar.axelarnet.v1beta1.Params"></a>
 
 ### Params
@@ -790,6 +807,7 @@ Params represent the genesis parameters for the module
 | `route_timeout_window` | [uint64](#uint64) |  | IBC packet route timeout window |
 | `transfer_limit` | [uint64](#uint64) |  |  |
 | `end_blocker_limit` | [uint64](#uint64) |  |  |
+| `call_contracts_proposal_min_deposits` | [CallContractProposalMinDeposit](#axelar.axelarnet.v1beta1.CallContractProposalMinDeposit) | repeated |  |
 
 
 
@@ -5811,23 +5829,6 @@ Query defines the gRPC querier service.
 
 
 
-<a name="axelar.nexus.v1beta1.CallContractProposalMinDeposit"></a>
-
-### CallContractProposalMinDeposit
-
-
-
-| Field | Type | Label | Description |
-| ----- | ---- | ----- | ----------- |
-| `chain` | [string](#string) |  |  |
-| `contract_address` | [string](#string) |  |  |
-| `min_deposits` | [cosmos.base.v1beta1.Coin](#cosmos.base.v1beta1.Coin) | repeated |  |
-
-
-
-
-
-
 <a name="axelar.nexus.v1beta1.Params"></a>
 
 ### Params
@@ -5840,7 +5841,6 @@ Params represent the genesis parameters for the module
 | `chain_maintainer_missing_vote_threshold` | [axelar.utils.v1beta1.Threshold](#axelar.utils.v1beta1.Threshold) |  |  |
 | `chain_maintainer_incorrect_vote_threshold` | [axelar.utils.v1beta1.Threshold](#axelar.utils.v1beta1.Threshold) |  |  |
 | `chain_maintainer_check_window` | [int32](#int32) |  |  |
-| `call_contracts_proposal_min_deposits` | [CallContractProposalMinDeposit](#axelar.nexus.v1beta1.CallContractProposalMinDeposit) | repeated |  |
