+++ proto/axelar/axelarnet/v1beta1/params.proto
@@ -3,6 +3,7 @@ package axelar.axelarnet.v1beta1;
 option go_package = "github.com/axelarnetwork/axelar-core/x/axelarnet/types";
 
 import "gogoproto/gogo.proto";
+import "cosmos/base/v1beta1/coin.proto";
 
 option (gogoproto.goproto_getters_all) = false;
 
@@ -13,4 +14,20 @@ message Params {
   reserved 2; // transaction_fee_rate was removed in v0.15
   uint64 transfer_limit = 3;
   uint64 end_blocker_limit = 4;
+  repeated CallContractProposalMinDeposit call_contracts_proposal_min_deposits =
+      5 [
+        (gogoproto.castrepeated) = "CallContractProposalMinDeposits",
+        (gogoproto.nullable) = false
+      ];
+}
+
+message CallContractProposalMinDeposit {
+  string chain = 1
+      [ (gogoproto.casttype) =
+            "github.com/axelarnetwork/axelar-core/x/nexus/exported.ChainName" ];
+  string contract_address = 2;
+  repeated cosmos.base.v1beta1.Coin min_deposits = 3 [
+    (gogoproto.castrepeated) = "github.com/cosmos/cosmos-sdk/types.Coins",
+    (gogoproto.nullable) = false
+  ];
 }
