+++ x/axelarnet/keeper/genesis.go
@@ -17,7 +17,7 @@ import (
 
 // InitGenesis initializes the axelarnet module's state from a given genesis state.
 func (k Keeper) InitGenesis(ctx sdk.Context, genState *types.GenesisState) {
-	k.setParams(ctx, genState.Params)
+	k.SetParams(ctx, genState.Params)
 	if len(genState.CollectorAddress) > 0 {
 		if err := k.SetFeeCollector(ctx, genState.CollectorAddress); err != nil {
 			panic(err)
@@ -61,7 +61,7 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 	collector, _ := k.GetFeeCollector(ctx)
 
 	return types.NewGenesisState(
-		k.getParams(ctx),
+		k.GetParams(ctx),
 		collector,
 		k.getCosmosChains(ctx),
 		k.GetIBCTransferQueue(ctx).(utils.GeneralKVQueue).ExportState(),
