+++ x/nexus/types/codec.go
@@ -5,7 +5,6 @@ import (
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
 // RegisterLegacyAminoCodec registers concrete types on codec
@@ -27,10 +26,6 @@ func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
 		&DeactivateChainRequest{},
 		&RegisterAssetFeeRequest{},
 		&SetTransferRateLimitRequest{},
-	)
-
-	registry.RegisterImplementations((*govtypes.Content)(nil),
-		&CallContractsProposal{},
 	)
 }
